fix: move DEFAULT_APP_NAME to lib/constants to avoid server imports in client components
Client components importing from lib/system-settings pulled in lib/db
(postgres), which requires Node.js fs module — breaks the browser
bundle. Move the constant to lib/constants.ts (no server deps) and
import from there in client components.
