+++ components/brand.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import { Container } from "lucide-react";
-import { DEFAULT_APP_NAME } from "@/lib/system-settings";
+import { DEFAULT_APP_NAME } from "@/lib/constants";
 
 const appName = process.env.NEXT_PUBLIC_APP_NAME || DEFAULT_APP_NAME;
