+++ lib/constants.ts
@@ -1,0 +1,2 @@
+/** Default app name used across the UI, emails, and metadata when not configured. */
+export const DEFAULT_APP_NAME = "Vardo";
